Re-namespace and apply suggested refactorings
