+++ osu.Framework.Templates/templates/template-flappy/FlappyDon.Game.Tests/Visual/TestSceneBackdrop.cs
@@ -1,3 +1,4 @@
+using FlappyDon.Game.Elements;
 using osu.Framework.Allocation;
 using osu.Framework.Testing;
