+++ osu.Framework.Templates/templates/template-flappy/FlappyDon.Game/Elements/Backdrop.cs
@@ -6,7 +6,7 @@ using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Sprites;
 using osuTK;
 
-namespace FlappyDon.Game
+namespace FlappyDon.Game.Elements
 {
     /// <summary>
     /// Manages a pool of Sprite objects,
