+++ osu.Framework.Templates/templates/template-flappy/FlappyDon.Game/Elements/Pipe.cs
@@ -4,7 +4,7 @@ using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.Textures;
 using osuTK;
 
-namespace FlappyDon.Game
+namespace FlappyDon.Game.Elements
 {
     /// <summary>
     /// A sprite that shows a single green pipe.
