+++ osu.Framework.Templates/templates/template-flappy/FlappyDon.Game/FlappyDonGame.cs
@@ -1,4 +1,5 @@
-﻿using osu.Framework.Allocation;
+﻿using FlappyDon.Game.Elements;
+using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Audio;
 using osu.Framework.Graphics.Containers;
